services/api/pyproject.toml
[tool.poetry]
package-mode = false
name = "api"
version = "0.1.0"
description = ""
authors = ["Renja <renja.grotemeyer@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
db_models = {path = "../../libs/db_models/"} # add "develop = true" when developing
pulsefire = "^2.0.8"
orjson = "^3.9.15"
python-dotenv = "^1.0.1"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic-settings = "^2.2.1"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"